refactor a few things, fix some bugs
first commit after tweaking the triangle shape in the resolv library. added a subpixel concept to moveable actors, fixed some issues introduced when making large triangles (depth was calculated wrong for the x and y axes), some overall tweaks

+++ actors/prefab_events.go
@@ -11,7 +11,7 @@ import (
 // ==== Actor Events
 
 // NewMoveByEvent w
-func NewMoveByEvent(target Actor, dx, dy, dz int, duration types.Frame) *events.Event {
+func NewMoveByEvent(target Actor, dx, dy, dz float64, duration types.Frame) *events.Event {
 	return events.New(events.Actor, MoveByActionType, []interface{}{target, dx, dy, dz, duration})
 }
